refactor: migrate GQL parser unit tests from tinkergraph-gremlin to gql-gremlin
QueryGraphTest and GqlGrammarTest had no TinkerGraph dependencies — they tested
pure MATCH parsing logic that lives in gql-gremlin. Moving them alongside the
code they test. The existing four error-message-format tests are merged into
the comprehensive QueryGraphTest that was living in the wrong module.
